refactor: BlueprintService와 ServiceImpl 클래스 통합
